refactor(T-05): Branch

+++ src/main/java/com/healthcare/backend/dto/response/BranchResponse.java
@@ -1,10 +1,15 @@
 package com.healthcare.backend.dto.response;
+
 import lombok.*;
-@Getter @Setter
-@NoArgsConstructor @AllArgsConstructor
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
 @Builder
 public class BranchResponse {
-    private Integer branchId;
+
+    private Long branchId;
     private String branchName;
     private String branchAddress;
     private String branchHotline;
